[MouseHighlighter]Update always-on size and color in real-time
Update always-on pointer size and color in real-time when changing them
while it's activated.

Recreate the pointer when the size has changed.

Update its color only while changing color for better UX.
This is because recreating the pointer while changing color caused
the pointer disappears for a moment in my test. I think the recreation
is slow and it's called so many times while changing color.